Cut dependency between many extensions and the core custom elements class.
This pulled in a lot of unneeded code leading to major size regression.
